Further refined Level 1, Fixed Urban Background
So I doubled the length of the background, and have done my best to seam it and make it flow as one long background. I defer to you as to whether I have succeeded, and or whether we want to put this into photoshop and scale it up there instead of in Unity, so that we can retain the quality of the pattern.